1.新增了摄像头功能 2.修复了一些BUG 3.调整了侧边栏项目

+++ MyE/Device/Socket/MyESocketManualViewController.m
@@ -27,7 +27,7 @@
     [super viewWillAppear:YES];
     if (self.needRefresh) {
         self.needRefresh = NO;
-        [self uploadOrDownloadInfoFromServerWithURL:[NSString stringWithFormat:@"%@?houseId=%i&tId=%@",GetRequst(URL_FOR_SOCKET_FIND),MainDelegate.houseData.houseId,self.device.tid] andName:@"downloadInfo"];
+        [self uploadOrDownloadInfoFromServerWithURL:[NSString stringWithFormat:@"%@?houseId=%li&tId=%@",GetRequst(URL_FOR_SOCKET_FIND),(long)MainDelegate.houseData.houseId,self.device.tid] andName:@"downloadInfo"];
     }
 }
 - (void)viewDidLoad
